parts: Move font.h for hd44780 display to parts directory
This file is referenced only by hd44780_glut.c in the parts folder,
so move it there and give it a prefix.
